Warn if too many/few arguments are passed to sprintf, fix bug
Fixes #2481
