+++ tests/misc/fallback_test/expected/064_invalid_constant_expression.php.expected
@@ -1,0 +1,1 @@
+src/064_invalid_constant_expression.php:2 PhanInvalidConstantExpression Constant expression contains invalid operations
