+++ tests/misc/fallback_test/src/064_invalid_constant_expression.php
@@ -1,0 +1,4 @@
+<?php
+function test_invalid_constant($x = function() {}) {
+    var_export($x);
+}
